% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMM_Binary.r
\name{CoMM_Binary_Testing}
\alias{CoMM_Binary_Testing}
\title{CoMM Binary}
\usage{
CoMM_Binary_Testing(Y, Z, X_1, X_2)
}
\arguments{
\item{Y}{trait vector.}

\item{Z}{gene expression vector.}

\item{X_1}{normalized genotype (cis-SNPs) matrix for eQTL.}

\item{X_2}{normalized genotype (cis-SNPs) matrix for GWAS.}
}
\value{
List of model parameters
}
\description{
CoMM to dissecting genetic contributions to complex traits by leveraging regulatory information.
}
\author{
Boran Gao, \email{borang@umich.edu}
}
