\name{CommBinary-package}
\alias{CommBinary-package}
\alias{CommBinary}
\docType{package}
\title{CoMM Binary 
\packageTitle{CommBinary}
}
\description{CoMM binary to dissecting genetic contributions to complex Disease by leveraging regulatory information.
\packageDescription{CommBinary}
}
\details{
CoMM_Binary_Testing is used for testing genetic contribution to complex diseases
\packageDESCRIPTION{CommBinary}
\packageIndices{CommBinary}
CoMM_Binary_Testing
}
\author{
\packageAuthor{CommBinary}
Boran Gao
Maintainer: \packageMaintainer{CommBinary} Boran Gao
}
