\name{CommBinary-package}
\alias{CommBinary-package}
\alias{CommBinary}
\docType{package}
\title{CoMM Binary 
\packageTitle{CommBinary}
}
\description{
\packageDescription{CommBinary}
}
\details{
Function CoMM_Binary_Testing is used for testing genetic contribution to complex diseases
}
\author{ 
\packageAuthor{CommBinary}

Maintainer: \packageMaintainer{CommBinary} 
}
